fix: remove participants and actions from post details view
