+++ src/screens/Posts/PostDetailsStack/PostSpecifics/PostSpecifics.tsx
@@ -1,10 +1,8 @@
 import { useEffect, useState } from "react";
 import { View } from "react-native";
 
-import PostDetailsActions from "./PostDetailsActions/PostDetailsActions";
 import PostDetailsDescription from "./PostDetailsDescription/PostDetailsDescription";
 import PostDetailsImage from "./PostDetailsImage/PostDetailsImage";
-import PostDetailsParticipants from "./PostDetailsParticipants/PostDetailsParticipants";
 import PostDetailsReactions from "./PostDetailsReactions/PostDetailsReactions";
 import { usePostDetailsStore } from "../../../../zustand/post-details";
 
@@ -32,8 +30,6 @@ const PostSpecifics = ({ route, navigation }: PostSpecificsProps) => {
         imageURI={imageURI}
       />
       <View style={{ padding: 10 }}>
-        <PostDetailsActions />
-        <PostDetailsParticipants />
         <PostDetailsDescription />
         <PostDetailsReactions />
       </View>
